Extract yorkieServer and rename clusterServer
